+++ src/announcements/editor.css
@@ -205,7 +205,7 @@
     margin-bottom: 15px;
 }
 
-.news-article {
+.announcements-article {
     width: 100%;
 }
