+++ src/i18n/ja/index.ts
@@ -3,7 +3,7 @@ import sidebar from './sidebar';
 import device from './device';
 import auth from './auth';
 import users from './users';
-import news from './news';
+import announcements from './announcements';
 
 export default {
   translation: {
@@ -12,6 +12,6 @@ export default {
     device,
     auth,
     users,
-    news
+    announcements
   },
 };
